requirements.txt
asgiref==3.10.0
asttokens==3.0.0
decorator==5.2.1
Django==5.2.8
django-debug-toolbar==6.1.0
django-extensions==4.1
executing==2.2.1
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pygraphviz==1.14
pyparsing==3.2.5
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.14
